aihero/project/pyproject.toml
[project]
name = "tech-interview-ai-assistant"
version = "0.1.0"
description = "AI Assistant for Tech Interview Preparation"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "streamlit>=1.32.0",
    "pydantic-ai>=0.0.14",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "minsearch>=0.0.5",
    "python-frontmatter>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]